feat: visionOS support (#5783)

+++ lib/services/ios-project-service.ts
@@ -63,14 +63,32 @@ interface INativeSourceCodeGroup {
 	files: string[];
 }
 
-const DevicePlatformSdkName = "iphoneos";
-const SimulatorPlatformSdkName = "iphonesimulator";
+export const DevicePlatformSdkName = "iphoneos";
+export const SimulatorPlatformSdkName = "iphonesimulator";
+export const VisionDevicePlatformSdkName = "xros";
+export const VisionSimulatorPlatformSdkName = "xrsimulator";
+
 const FRAMEWORK_EXTENSIONS = [".framework", ".xcframework"];
 
-const getPlatformSdkName = (forDevice: boolean): string =>
-	forDevice ? DevicePlatformSdkName : SimulatorPlatformSdkName;
-const getConfigurationName = (release: boolean): string =>
-	release ? Configurations.Release : Configurations.Debug;
+const getPlatformSdkName = (buildData: IBuildData): string => {
+	const forDevice =
+		!buildData || buildData.buildForDevice || buildData.buildForAppStore;
+	const isvisionOS = injector
+		.resolve("devicePlatformsConstants")
+		.isvisionOS(buildData.platform);
+
+	if (isvisionOS) {
+		return forDevice
+			? VisionDevicePlatformSdkName
+			: VisionSimulatorPlatformSdkName;
+	}
+
+	return forDevice ? DevicePlatformSdkName : SimulatorPlatformSdkName;
+};
+
+const getConfigurationName = (release: boolean): string => {
+	return release ? Configurations.Release : Configurations.Debug;
+};
 
 export class IOSProjectService extends projectServiceBaseLib.PlatformProjectServiceBase {
 	private static IOS_PROJECT_NAME_PLACEHOLDER = "__PROJECT_NAME__";
@@ -78,6 +96,7 @@ export class IOSProjectService extends projectServiceBaseLib.PlatformProjectServ
 
 	constructor(
 		$fs: IFileSystem,
+		private $options: IOptions,
 		private $childProcess: IChildProcess,
 		private $cocoapodsService: ICocoaPodsService,
 		private $errors: IErrors,
@@ -101,7 +120,8 @@ export class IOSProjectService extends projectServiceBaseLib.PlatformProjectServ
 		private $iOSNativeTargetService: IIOSNativeTargetService,
 		private $sysInfo: ISysInfo,
 		private $tempService: ITempService,
-		private $spmService: ISPMService
+		private $spmService: ISPMService,
+		private $mobileHelper: Mobile.IMobileHelper
 	) {
 		super($fs, $projectDataService);
 	}
@@ -120,19 +140,22 @@ export class IOSProjectService extends projectServiceBaseLib.PlatformProjectServ
 			projectData.platformsDir &&
 			this._platformsDirCache !== projectData.platformsDir
 		) {
+			const platform = this.$mobileHelper.normalizePlatformName(
+				this.$options.platformOverride ?? this.$devicePlatformsConstants.iOS
+			);
 			const projectRoot = path.join(
 				projectData.platformsDir,
-				this.$devicePlatformsConstants.iOS.toLowerCase()
+				platform.toLowerCase()
 			);
 			const runtimePackage = this.$projectDataService.getRuntimePackage(
 				projectData.projectDir,
-				constants.PlatformTypes.ios
+				platform.toLowerCase() as constants.SupportedPlatform
 			);
 
 			this._platformData = {
 				frameworkPackageName: runtimePackage.name,
-				normalizedPlatformName: "iOS",
-				platformNameLowerCase: "ios",
+				normalizedPlatformName: platform,
+				platformNameLowerCase: platform.toLowerCase(),
 				appDestinationDirectoryPath: path.join(
 					projectRoot,
 					projectData.projectName
@@ -144,9 +167,7 @@ export class IOSProjectService extends projectServiceBaseLib.PlatformProjectServ
 					return path.join(
 						projectRoot,
 						constants.BUILD_DIR,
-						`${config}-${getPlatformSdkName(
-							!options || options.buildForDevice || options.buildForAppStore
-						)}`
+						`${config}-${getPlatformSdkName(options)}`
 					);
 				},
 				getValidBuildOutputData: (
@@ -654,11 +675,19 @@ export class IOSProjectService extends projectServiceBaseLib.PlatformProjectServ
 
 			this.savePbxProj(project, projectData);
 
-			const resourcesNativeCodePath = path.join(
+			let resourcesNativeCodePath = path.join(
 				resourcesDirectoryPath,
 				platformData.normalizedPlatformName,
 				constants.NATIVE_SOURCE_FOLDER
 			);
+
+			if (!this.$fs.exists(resourcesNativeCodePath)) {
+				resourcesNativeCodePath = path.join(
+					resourcesDirectoryPath,
+					this.$devicePlatformsConstants.iOS,
+					constants.NATIVE_SOURCE_FOLDER
+				);
+			}
 
 			await this.prepareNativeSourceCode(
 				constants.TNS_NATIVE_SOURCE_GROUP_NAME,
@@ -696,14 +725,28 @@ export class IOSProjectService extends projectServiceBaseLib.PlatformProjectServ
 		this.$fs.deleteDirectory(platformsAppResourcesPath);
 		this.$fs.ensureDirectoryExists(platformsAppResourcesPath);
 
-		this.$fs.copyFile(
-			path.join(
-				projectAppResourcesPath,
-				platformData.normalizedPlatformName,
-				"*"
-			),
-			platformsAppResourcesPath
+		const platformAppResourcesPath = path.join(
+			projectAppResourcesPath,
+			platformData.normalizedPlatformName
 		);
+
+		// this allows App_Resources/visionOS
+		if (this.$fs.exists(platformAppResourcesPath)) {
+			this.$fs.copyFile(
+				path.join(platformAppResourcesPath, "*"),
+				platformsAppResourcesPath
+			);
+		} else {
+			// otherwise falls back to App_Resources/iOS
+			this.$fs.copyFile(
+				path.join(
+					projectAppResourcesPath,
+					this.$devicePlatformsConstants.iOS,
+					"*"
+				),
+				platformsAppResourcesPath
+			);
+		}
 
 		this.$fs.deleteFile(
 			path.join(platformsAppResourcesPath, platformData.configurationFileName)
